Jan 17, 2025, 10:43 AM

+++ CT06_02/lesson2.py
@@ -38,5 +38,5 @@ print(colours_of_the_rainbow)
 test1 = (input("What is your WA1 score?"))
 test2 = (input("What is your WA2 score?"))
 test3 = (input("What is your WA3 score?"))
-final_score = 0.2*test1 + 0.4*test2 + 0.4*test3
+final_score = float(0.2*test1 + 0.4*test2 + 0.4*test3)
 print("Your final score is" + final_score + ".")
